Add new_position function to calculate position given old position, velocity and elapsed time. Change components from f32 to f64 representations. (#1)

+++ stacktrader-types/src/components.rs
@@ -1,9 +1,9 @@
 /// Represents a position in 3-dimensional space, assumed unit is Kilometers
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 /// Represents a velocity, which includes a magnitude and a direction. The direction
@@ -11,7 +11,7 @@ pub struct Position {
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct Velocity {
     pub mag: u32,
-    pub ux: f32,
-    pub uy: f32,
-    pub uz: f32,
+    pub ux: f64,
+    pub uy: f64,
+    pub uz: f64,
 }
